Refactor code style, remove multiple roles
Initial intent of multiple "CREATE ROLE" was to perform multiple
connections to the database for different account types. Spring Boot
does not support multiple connections to the same database, unless a
microservice approach is adopted (overkill).

All queries are consolidated into a single file, schema.sql, with test
queries placed in the test folder. Queries within schema.sql will be
used in deployment, i.e. registering for a cloud database provider,
executing the schema.sql on the instance of postgres in the cloud.

JSON and "CREATE FUNCTION" will likely not be used, as Prof. remarked
that our primary backend logic should be done in the backend framework
as opposed to the database.

An example api is implemented at "/hello", which outputs hello in the
browser or curl GET request. Clone the repository to test it out.