Update schema, add controller examples
The schema is now updated to our latest user stories. See commit date.
The controller, i.e. the api features a few examples on creating an
account user randomly generated data. The 'backend version' of the
database schema can be seen in main/com.hotdog.ctbs/entity/UserAccount
A skeleton for the first service is created in the service folder,
this is subject to change. We will implement the design of using a
service if we have time. The service will be used to handle the logic
of the application, i.e. the business logic. The controller will be
used to handle the api calls. Newly implemented changes may break the
execution of running the application. For now, review the schema and
controller examples. Ask questions if you have any.