Update schema and docker, add controller examples
Executing the script `./schema.sh` will start a container and load the
schema. If a container already exists, a new container will replace the
existing one.

Code styles for this project can be synced with project files from the
.idea folder. Please use IntelliJ to maintain the code style.

Entity for user profile has been added. See the GET and POST requests
in controller/admin/CreateUserProfile.java and tests in the test file
test/.../ctbs/CinemaTicket...Tests.java

Implementation and fine-tuning of each CREATE TABLE will be done in
the order they were listed in on the schema.sql file. All reasoning
and context will also be commented, this is our SSOT.