Update schema and UserProfile, refactor code
Implementation, controller, and tests for UserProfile are done. All
UserProfileImpl methods are tested in UserProfileTests. Methods from
UserProfile, UserProfileRepository, and UserProfileController are
kept bare, to centralize the logic in one file (UserProfileImpl).

Some testing files were deleted, and code styles are being changed to
improve readability. To test the implementation, click the play button
next to each method in UserProfileTests.